+++ src/main/java/nl/tudelft/skills/service/SkillService.java
@@ -17,9 +17,28 @@
  */
 package nl.tudelft.skills.service;
 
+import static java.util.stream.Collectors.toMap;
+
+import java.util.ArrayList;
+import java.util.Comparator;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
+import java.util.Stack;
+import java.util.function.Function;
+import java.util.stream.Collectors;
+
+import nl.tudelft.labracore.api.CourseControllerApi;
+import nl.tudelft.labracore.api.EditionControllerApi;
+import nl.tudelft.labracore.api.dto.CourseDetailsDTO;
+import nl.tudelft.labracore.api.dto.EditionSummaryDTO;
 import nl.tudelft.skills.model.AbstractSkill;
+import nl.tudelft.skills.model.ExternalSkill;
 import nl.tudelft.skills.model.Skill;
 import nl.tudelft.skills.repository.AbstractSkillRepository;
+import nl.tudelft.skills.repository.EditionRepository;
+import nl.tudelft.skills.repository.SkillRepository;
 import nl.tudelft.skills.repository.TaskCompletionRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -31,12 +50,22 @@ public class SkillService {
 
 	private final AbstractSkillRepository abstractSkillRepository;
 	private final TaskCompletionRepository taskCompletionRepository;
+	private final EditionRepository editionRepository;
+	private final EditionControllerApi editionApi;
+	private final CourseControllerApi courseApi;
+	private final SkillRepository skillRepository;
 
 	@Autowired
 	public SkillService(AbstractSkillRepository abstractSkillRepository,
-			TaskCompletionRepository taskCompletionRepository) {
+			TaskCompletionRepository taskCompletionRepository, EditionControllerApi editionApi,
+			CourseControllerApi courseApi, SkillRepository skillRepository,
+			EditionRepository editionRepository) {
 		this.abstractSkillRepository = abstractSkillRepository;
 		this.taskCompletionRepository = taskCompletionRepository;
+		this.editionApi = editionApi;
+		this.courseApi = courseApi;
+		this.skillRepository = skillRepository;
+		this.editionRepository = editionRepository;
 	}
 
 	/**
@@ -51,9 +80,97 @@ public class SkillService {
 		skill.getChildren().forEach(c -> c.getParents().remove(skill));
 		if (skill instanceof Skill s) {
 			s.getTasks().forEach(t -> taskCompletionRepository.deleteAll(t.getCompletedBy()));
+			s.getFutureEditionSkills().forEach(innerSkill -> innerSkill.setPreviousEditionSkill(null));
+
+			if (s.getPreviousEditionSkill() != null) {
+				s.getPreviousEditionSkill().getFutureEditionSkills().remove(s);
+			}
 		}
 		abstractSkillRepository.delete(skill);
 		return skill;
+	}
+
+	/**
+	 * Gets the correct skill for an external skill. This is the skill in the most recent edition which the
+	 * person has last worked on. If none such edition exists, the most recent edition is chosen. "Most
+	 * recent" refers to the start date of the edition, and not the date of activity by the person.
+	 *
+	 * @param  personId      The id of the person.
+	 * @param  externalSkill The external skill.
+	 * @return               The correct skill which the external skill refers to. This is the skill in the
+	 *                       most recent edition which the person has last worked on. If none such edition
+	 *                       exists, the most recent edition is chosen.
+	 */
+	public Skill recentActiveEditionForSkillOrLatest(Long personId, ExternalSkill externalSkill) {
+		Skill skill = externalSkill.getSkill();
+		Long editionId = skill.getSubmodule().getModule().getEdition().getId();
+		CourseDetailsDTO course = courseApi.getCourseByEdition(editionId).block();
+
+		Map<Long, EditionSummaryDTO> editionsById = course.getEditions().stream()
+				.collect(toMap(EditionSummaryDTO::getId, Function.identity()));
+
+		// Get ids of the editions the person as completed at least one task in
+		Set<Long> completedTasksInEditions = taskCompletionRepository.getByPersonId(personId).stream()
+				.map(taskCompletion -> taskCompletion.getTask().getSkill().getSubmodule()
+						.getModule().getEdition().getId())
+				.collect(Collectors.toSet());
+
+		// Do DFS
+		List<Skill> traversal = traverseSkillTree(skill);
+
+		// Filter by skills in visible editions
+		traversal = traversal.stream()
+				.filter(innerSkill -> {
+					Long innerEditionId = innerSkill.getSubmodule().getModule().getEdition().getId();
+					return editionRepository.getById(innerEditionId).isVisible();
+				})
+				.sorted(Comparator.comparing((Skill innerSkill) -> editionsById.get(innerSkill.getSubmodule()
+						.getModule().getEdition().getId()).getStartDate()).reversed())
+				.collect(Collectors.toList());
+
+		// If it exists, return the skill from the last edition the person has completed a task in
+		Optional<Skill> completedTasksInEdition = traversal.stream()
+				.filter(innerSkill -> completedTasksInEditions.contains(innerSkill.getSubmodule()
+						.getModule().getEdition().getId()))
+				.findFirst();
+
+		// If it does not exist, return the skill in the latest edition
+		// Traversal should never be empty, as it at least contains the initial skill
+		return completedTasksInEdition.orElse(traversal.get(0));
+	}
+
+	/**
+	 * Return the DFS traversal of the "skill tree" for the given tree. The tree contains all skill versions
+	 * in different editions, linked by the previous/future skill(s) attributes
+	 *
+	 * @param  skill The skill for which to traverse the tree of skill versions.
+	 * @return       The DFS traversal of the skill tree for the given skill.
+	 */
+	public List<Skill> traverseSkillTree(Skill skill) {
+		// Find root of the "skill tree"
+		Skill current = skill;
+		while (current.getPreviousEditionSkill() != null) {
+			current = current.getPreviousEditionSkill();
+		}
+
+		// Do DFS
+		Stack<Skill> traversal = new Stack<>();
+		traversal.push(current);
+
+		List<Skill> skills = new ArrayList<>();
+
+		while (!traversal.isEmpty()) {
+			current = traversal.pop();
+			skills.add(current);
+
+			// Continue traversal
+			Set<Skill> nextSkills = current.getFutureEditionSkills();
+			for (Skill nextSkill : nextSkills) {
+				traversal.push(nextSkill);
+			}
+		}
+
+		return skills;
 	}
 
 }